+++ schemas/schema.ts
@@ -14,8 +14,8 @@ import localeBlock from './localeBlock';
 import localeString from './localeString';
 import dinLivssituasjonDokumenter from './steg/din-livssituasjon/dinLivssituasjonDokumenter';
 import dokumentasjonDokumenter from './steg/dokumentasjon/dokumentasjonDokumenter';
-import eosForBarnDokumenter from './steg/eosForBarn/eosForBarnDokumenter';
-import eosForSokerDokumenter from './steg/eosForSoker/eosForSokerDokumenter';
+import eosForBarnDokumenter from './steg/eos-for-barn/eosForBarnDokumenter';
+import eosForSokerDokumenter from './steg/eos-for-soker/eosForSokerDokumenter';
 import forsideDokumenter from './steg/forside/forsideDokumenter';
 import kvitteringDokumenter from './steg/kvittering/kvitteringDokumenter';
 import omBarnaDokumenter from './steg/om-barna/omBarnaDokumenter';
