+++ schemas/steg/eos-for-barn/tittel.ts
