+++ schemas/steg/eos-for-soker/eosForSokerDokumenter.ts
