+++ schemas/steg/eos-for-soker/tittel.ts
